Bunch of (build breaking) stuff. This is temporary.
design diagrams for widgets and entity system
Updates for act library for C++11 types
Adding components
methods for 4x4 matrix
constructors for Rect
window and sizer base classes
moving display timer

+++ src/agt/agta/agta_entity.cpp
@@ -2,7 +2,7 @@
 
 namespace agta {
 
-Entity::Entity(uint32_t id, std::string const& name)
+Entity::Entity(size_t id, std::string const& name)
 : m_id(id),
   m_name(name)
 {}
@@ -10,7 +10,7 @@ Entity::Entity(uint32_t id, std::string const& name)
 Entity::~Entity()
 {}
 
-uint32_t Entity::id() const
+size_t Entity::id() const
 {
     return m_id;
 }
